Moved source/utils to source/Grid/GridUtils for better organization
